+++ src/lib.rs
@@ -9,8 +9,6 @@
 )]
 
 mod client;
-mod socket;
-mod tcp;
 
 pub use client::Builder;
 pub use client::Client;